requirements_test.txt
asynctest==0.9.0
flake8>=3.0.4
pylint>=1.5.6
coveralls>=1.1
pytest>=2.9.2
pytest-cov>=2.3.1
pytest-timeout>=1.0.0
pytest-aiohttp==0.1.3
pydocstyle>=1.0.0
mypy-lang>=0.4
Sphinx==1.5.1
